pull inline chat zone widget into its own file
